fix(diagnostics): use monotonic cursor for stderr sampling + biome format
Switch sample_stderr_logs to use get_logs_since(cursor: u64) instead of
length-based tracking. The old approach stalled when the ring buffer was
at steady-state (evict-one-per-append) and caused duplicate LogPoint
entries on reset after eviction.

Also fix TraceExplorer.tsx line length so biome format check passes.

+++ crates/logfwd-diagnostics/src/diagnostics/server.rs
@@ -1138,7 +1138,7 @@ async fn handle_ws(mut socket: WebSocket, state: Arc<DiagnosticsState>) {
 // ---------------------------------------------------------------------------
 
 async fn sampler_loop(state: Arc<DiagnosticsState>) {
-    let mut last_stderr_cursor: usize = 0;
+    let mut last_stderr_cursor: u64 = 0;
     let mut prev_health = std::collections::HashMap::new();
     let mut prev_snapshot: Option<super::telemetry::MetricSnapshot> = None;
     loop {
